as_vevent: use vDDDTypes instead of explicit date/datetime type
This fixes a problem when using the `decoded` method.
E.g., `event.decoded('dtstart')` fails with an exception:

```
  File ".../.pyenv/lib/python3.7/site-packages/icalendar/cal.py", line 237, in decoded
    return self._decode(name, value)
  File ".../.pyenv/lib/python3.7/site-packages/icalendar/cal.py", line 220, in _decode
    decoded = types_factory.from_ical(name, value)
  File ".../.pyenv/lib/python3.7/site-packages/icalendar/prop.py", line 1042, in from_ical
    decoded = type_class.from_ical(value)
  File ".../.pyenv/lib/python3.7/site-packages/icalendar/prop.py", line 334, in from_ical
    u = ical.upper()
```

This is due to a type inconsistency between the event created by this
package and the type that `icalendar` expects.

Another (ugly) workaround for this problem (if this patch was not
applied) would be to convert the event to ical and parse it back using
`icalendar`, which repairs the dtstart type inconsistency as well (e.g.,
something like `event = icalendar.Calendar.from_ical(event.to_ical())``