Remove pytz dependency
This changes the types of timezones that are accepted by the library,
from pytz instances to those that satisfy PEP-495.

On Python 3.9 and above, the `zoneinfo` built-in module provides a
compatible class.  On Python 3.6 and above, `backports.zoneinfo` can
be used in its place, or `dateutil` provides its own implementation.
Users can also continue to use `pytz` timezones by modifying their code
to wrap them with `pytz_deprecation_shim` (this change could be extended
to apply this shim within the library itself, avoiding the need for
users to change their code).

Python versions below 3.6 will no longer be supported (due to the use of
`datetime.astimezone()`), though the major Linux distributions all carry
3.6+, and upstream support for older versions has long since ended.